Inline all js/css dependencies into the bundle
Removes the need for the cssify gulp task (which was building css separetely to the js) as well as simplifying the index.html file. It's now needing to only load the bundle.js file in to start the game.

We're also now loading jquery from npm, instead of holding a copy in the repository. I would like to be able to do this with Semantic UI as well however there was some error when trying to install it. I will work on this another time....